Add condition for patch with version only
RISDEV-5750

* Adding test for patch update with only version
* Add contain only version method
* Add test for contain only patch method

+++ backend/src/main/java/de/bund/digitalservice/ris/caselaw/domain/DocumentationUnitService.java
@@ -287,7 +287,8 @@ public class DocumentationUnitService {
 
     JsonPatch toFrontendJsonPatch = new JsonPatch(Collections.emptyList());
     RisJsonPatch toFrontend;
-    if (!patch.patch().getOperations().isEmpty()) {
+    if (!patch.patch().getOperations().isEmpty()
+        && !PatchMapperService.containsOnlyVersionInPatch(patch.patch())) {
       JsonPatch toUpdate = patchMapperService.removePatchForSamePath(patch.patch(), newPatch);
 
       log.debug("version {} - update patch: {}", patch.documentationUnitVersion(), toUpdate);
